+++ app/models/assessment.rb
@@ -16,7 +16,6 @@ class Assessment < ApplicationRecord
   validate :short_identifier_unchanged, on: :update
   validates_presence_of :description
   validates_inclusion_of :is_hidden, in: [true, false]
-  validates_presence_of :notes_count
   validates :short_identifier, format: { with: /\A[a-zA-Z0-9\-_]+\z/,
                                          message: 'short_identifier must only contain alphanumeric, hyphen, or '\
                                                   'underscore' }